Handle escape characters in the packet structure.
Chars 0x55, 0x3c and 0xc3 are special. The first begins a packet (and, I
assume would terminate parsing if encountered unexpectedly), the second
ends a packet, and the third is used to fix this.

Basically, 0xc3 escapes the next byte. It gets dropped itself from all
calculations and disables any special handling for 1 byte.